[Feat] 좋아요 캐시카운터에서 직접 테이블 count 반환

+++ src/main/java/com/swyp/app/domain/perspective/dto/response/LikeCountResponse.java
@@ -2,4 +2,4 @@ package com.swyp.app.domain.perspective.dto.response;
 
 import java.util.UUID;
 
-public record LikeCountResponse(UUID perspectiveId, int likeCount) {}
+public record LikeCountResponse(UUID perspectiveId, long likeCount) {}
